+++ regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/TRegexCompilationRequest.java
@@ -192,10 +192,10 @@ public final class TRegexCompilationRequest {
             if (pureNFA.getASTSubtree(lookAround).asLookAroundAssertion().isLiteral()) {
                 lookAroundExecutors[i] = new TRegexLiteralLookAroundExecutorNode(pureNFA.getASTSubtree(lookAround).asLookAroundAssertion(), compilationBuffer);
             } else {
-                lookAroundExecutors[i] = new TRegexBacktrackingNFAExecutorNode(pureNFA, lookAround, lookAroundExecutors, compilationBuffer);
+                lookAroundExecutors[i] = new TRegexBacktrackingNFAExecutorNode(pureNFA, lookAround, lookAroundExecutors, false, compilationBuffer);
             }
         }
-        return new TRegexBacktrackingNFAExecutorNode(pureNFA, pureNFA.getRoot(), lookAroundExecutors, compilationBuffer);
+        return new TRegexBacktrackingNFAExecutorNode(pureNFA, pureNFA.getRoot(), lookAroundExecutors, ast.getOptions().isMustAdvance(), compilationBuffer);
     }
 
     @TruffleBoundary
@@ -209,17 +209,19 @@ public final class TRegexCompilationRequest {
             preCalculatedResults = new PreCalculatedResultFactory[]{PreCalcResultVisitor.createResultFactory(ast)};
         }
         if (allowSimpleCG && preCalculatedResults == null && TRegexOptions.TRegexEnableTraceFinder && !ast.getRoot().hasLoops() && properties.isFixedCodePointWidth()) {
-            try {
-                phaseStart("TraceFinder NFA");
-                traceFinderNFA = NFATraceFinderGenerator.generateTraceFinder(nfa);
-                preCalculatedResults = traceFinderNFA.getPreCalculatedResults();
-                phaseEnd("TraceFinder NFA");
-                debugTraceFinder();
-            } catch (UnsupportedRegexException e) {
-                phaseEnd("TraceFinder NFA Bailout");
-                Loggers.LOG_BAILOUT_MESSAGES.fine(() -> "TraceFinder: " + e.getReason() + ": " + source);
-                // handle with capture group aware DFA, bailout will always happen before
-                // assigning preCalculatedResults
+            if (nfa.isFixedCodePointWidth()) {
+                try {
+                    phaseStart("TraceFinder NFA");
+                    traceFinderNFA = NFATraceFinderGenerator.generateTraceFinder(nfa);
+                    preCalculatedResults = traceFinderNFA.getPreCalculatedResults();
+                    phaseEnd("TraceFinder NFA");
+                    debugTraceFinder();
+                } catch (UnsupportedRegexException e) {
+                    phaseEnd("TraceFinder NFA Bailout");
+                    Loggers.LOG_BAILOUT_MESSAGES.fine(() -> "TraceFinder: " + e.getReason() + ": " + source);
+                    // handle with capture group aware DFA, bailout will always happen before
+                    // assigning preCalculatedResults
+                }
             }
         }
         boolean traceFinder = preCalculatedResults != null;